+++ asset/asset-factory/lib/index.js
@@ -1,0 +1,12 @@
+const { AssetStatus, AssetType } = require('./enum');
+const AssetIssuer = require('./issuer');
+const AssetRecipient = require('./recipient');
+const AssetFactory = require('./factory');
+
+module.exports = {
+  AssetStatus,
+  AssetType,
+  AssetIssuer,
+  AssetRecipient,
+  AssetFactory,
+};